+++ src/main/java/org/github/springframework/aop/BeforeAdvice.java
@@ -1,0 +1,8 @@
+package org.github.springframework.aop;
+import org.aopalliance.aop.Advice;
+/**
+ * @author zhaoyu
+ * @date 2024/1/27 17:20
+ */
+public interface BeforeAdvice extends Advice {
+}